# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mistral.rs",
    "description": "Fast, flexible LLM inference.",
    "contact": {
      "name": "Eric Buehler"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.9.4"
  },
  "paths": {
    "/calibration/apply": {
      "post": {
        "tags": [
          "Mistral.rs"
        ],
        "operationId": "calibration_apply",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalibrationApplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Requantize with collected statistics and hot-swap the layers.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalibrationStatus"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/agent/approvals/{approval_id}": {
      "post": {
        "tags": [
          "Mistral.rs"
        ],
        "operationId": "resolve_agent_approval",
        "parameters": [
          {
            "name": "approval_id",
            "in": "path",
            "description": "Approval ID from the approval-required SSE event",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Decision applied or queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalDecisionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid decision payload"
          },
          "404": {
            "description": "Unknown approval ID"
          },
          "413": {
            "description": "Decision payload is too large"
          },
          "415": {
            "description": "Unsupported content type"
          }
        }
      }
    },
    "/v1/audio/speech": {
      "post": {
        "tags": [
          "Mistral.rs"
        ],
        "summary": "Speech generation endpoint handler.",
        "operationId": "speech_generation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeechGenerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Speech generation"
          }
        }
      }
    },
    "/v1/chat/completions": {
      "post": {
        "tags": [
          "Mistral.rs"
        ],
        "summary": "OpenAI-compatible chat completions endpoint handler.",
        "operationId": "chatcompletions",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCompletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Chat completion JSON or server-sent event chunks",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatCompletionResponse"
                }
              },
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ChatCompletionChunkResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/completions": {
      "post": {
        "tags": [
          "Mistral.rs"
        ],
        "summary": "OpenAI-compatible completions endpoint handler.",
        "operationId": "completions",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Completion JSON or server-sent event chunks",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompletionResponse"
                }
              },
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/CompletionChunkResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/containers/{container_id}/files": {
      "get": {
        "tags": [
          "Mistral.rs"
        ],
        "operationId": "list_container_files",
        "parameters": [
          {
            "name": "container_id",
            "in": "path",
            "description": "Container ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of container file metadata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContainerFileMetadata"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/v1/containers/{container_id}/files/{file_id}": {
      "get": {
        "tags": [
          "Mistral.rs"
        ],
        "operationId": "get_container_file",
        "parameters": [
          {
            "name": "container_id",
            "in": "path",
            "description": "Container ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "file_id",
            "in": "path",
            "description": "File ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Container file metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContainerFileMetadata"
                }
              }
            }
          },
          "404": {
            "description": "File not found or expired"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/v1/containers/{container_id}/files/{file_id}/content": {
      "get": {
        "tags": [
          "Mistral.rs"
        ],
        "operationId": "get_container_file_content",
        "parameters": [
          {
            "name": "container_id",
            "in": "path",
            "description": "Container ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "file_id",
            "in": "path",
            "description": "File ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Raw file bytes with the file's MIME type"
          },
          "404": {
            "description": "File not found or expired"
          },
          "410": {
            "description": "File body was elided and is no longer fetchable"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/v1/embeddings": {
      "post": {
        "tags": [
          "Mistral.rs"
        ],
        "operationId": "embeddings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmbeddingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Embeddings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbeddingResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/files": {
      "get": {
        "tags": [
          "Mistral.rs"
        ],
        "operationId": "list_files",
        "responses": {
          "200": {
            "description": "List of file metadata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FileMetadata"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "Mistral.rs"
        ],
        "operationId": "upload_file",
        "requestBody": {
          "content": {
            "multipart/form-data": {}
          }
        },
        "responses": {
          "200": {
            "description": "Uploaded file metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileMetadata"
                }
              }
            }
          },
          "400": {
            "description": "Invalid upload"
          }
        }
      }
    },
...
    "/v1/files/{id}/content": {
      "get": {
        "tags": [
          "Mistral.rs"
        ],
        "operationId": "get_file_content",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "File ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Raw file bytes with the file's MIME type"
          },
          "404": {
            "description": "File not found or expired"
          },
          "410": {
            "description": "File body was elided and is no longer fetchable"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/v1/images/generations": {
      "post": {
        "tags": [
          "Mistral.rs"
        ],
        "summary": "Image generation endpoint handler.",
        "operationId": "image_generation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageGenerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Image generation"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApprovalDecisionResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "description": "\"resolved\" or \"queued\"."
          }
        }
      },
...
      "CalibrationApplyRequest": {
        "type": "object",
        "description": "Request body for applying online calibration.",
        "properties": {
          "save_cimatrix": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optionally save the collected imatrix to this `.cimatrix` file name in the server's working directory."
          }
        }
      },
      "CalibrationStatus": {
        "type": "object",
        "required": [
          "collecting",
          "layers",
          "layers_tracking",
          "total_rows",
          "min_rows",
          "max_rows"
        ],
        "properties": {
          "collecting": {
            "type": "boolean"
          },
          "layers": {
            "type": "integer",
            "minimum": 0
          },
          "layers_tracking": {
            "type": "integer",
            "minimum": 0
          },
          "max_rows": {
            "type": "integer",
            "minimum": 0
          },
          "min_rows": {
            "type": "integer",
            "minimum": 0
          },
          "total_rows": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
    }
  },
...
}